Added the Java version of the script
Updated the readme to explain how to use the Java version and to explain how the Python version was not tested properly.
The Java version only checks the command line inputs for now. Still updating.
Added an example DL.txt file. This file can be called anything but all the URLS should be on it's own line. Do not put any comments or other inputs into the file. It is only meant for URLs.

Signed-off-by: H1ddenGames <skarim3000@gmail.com>