Updated file structure, added youtube-dl method, updated command runner
Set the path to the executables in the script itself rather than relying on the user to have it in their path.
Created a youtubedl method that takes in a command and url. This method is reused in two ways:
1. It will get the url where the actual media is stored based on the url given through youtubedl's --get-url command.
2. It will get the 'human-readable' filename in ASCII with a limit of 260 characters through youtubedl's --get-filename -o '%(title)s.%(ext)s' --restrict-filenames command.

Updated readme to reflect the new file structure. (Moved Download.java from it's own intellij project to the main directory.)
Updated the command runner to use the same thread that the java script is running on. Later on, a new thread will be created to run the youtubedl commands in order then it will run the aria2c command to download.

Signed-off-by: H1ddenGames <skarim3000@gmail.com>