Added more tests and checks user input better, added comments
Updated the README to better explain how to use the script and what it does.
Updated the script with better comments.
The script now contains more tests and checks user input better.

The youtube-dl commands work properly but aria2c command still needs to be implemented.
Multithreading still needs to be implemented.

Signed-off-by: H1ddenGames <skarim3000@gmail.com>