Added SQLite and database functionality to the application
An SQLite database was implemented in order to keep track of files that were successfully downloaded and files that weren't able to be downloaded for various reasons (unable to find url, video no longer exists, command line error, etc)
sqlite-jdbc was included in order to interface with the SQLite database.
SQLiteTest was included for testing purposes. Once fully integrated with the application, it will be removed.

Both SQLiteTest and sqlite-jdbc files will be removed in a future update once a jar file is created with the final version of the application.

Signed-off-by: H1ddenGames <skarim3000@gmail.com>