Broke all functionality
Converted the project from unstructured to a Maven Project.
Restructured the directories for easier accessibility.
Will provide a jar file to run in order to use the Java version of the project.
The Python version of the project will be in the src/main/python folder.
The Python version needs to be updated for the new directory structure.

Added dependencies through the .pom file.

Signed-off-by: H1ddenGames <skarim3000@gmail.com>