Downloader is functional
The downloader is now able to correctly download a video when given a valid URL and able to display an error when given an invalid or unreachable URL.
The downloader still does not have multithreaded functionality.

The database is set up and able to accept inputs; however, the downloader does not currently use the database.

Order of importance:
1. Multithreaded downloads
2. Using the database to check if a video has been downloaded (to avoid redownloads and the cpu usage when youtube-dl is used to check a URL)

Signed-off-by: H1ddenGames <skarim3000@gmail.com>