Added dependencies, added stability
Added dependencies that allow maven to generate a jar and include sqlite jdbc dependency inside the jar.
Added nullpointerexception checks to make sure the script does not crash while running.
Changed the way executables are found so that packaging a jar is easier.
Updated code to check if running as a jar from the command line but kept functionality the same.
	- This means that the script will run the same whether or not the script is run using .java or .jar
Updated the documentation on how the script should be used.


Need to implement Multithreading according to the threads given in the input rather than creating as many threads as there are videos in the downloadFile.
	- Multithreading does work but not as intended. Might leave it how it currently is.
Need to clean up commented test code.
Need to implement database functionality/usage.
Need to test jar distribution -- running the java version works properly but the jar version sometimes gets stuck on downloads.


Signed-off-by: H1ddenGames <skarim3000@gmail.com>